user added to programs

+++ api/routers/program_router.py
@@ -1,5 +1,5 @@
 from typing import List
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Depends
 from sqlmodel import select, delete
 
 from api.database import SessionDep
@@ -12,7 +12,8 @@ from api.models.program_models import (
     ProgramUpdate,
 )
 from api.models.program_day_activities_link import ProgramDayActivityLink
-
+from api.security.auth import get_current_user # Importuj funkcję z Twojego pliku auth.py
+from api.models.user_models import Users
 # This file contains API endpoints related to Programs
 
 router = APIRouter(prefix="/api/programs", tags=["Programs"])
@@ -50,9 +51,11 @@ def link_days(session: SessionDep, program_id: int, days_in: List[ProgramDayInpu
 
 
 @router.post("", response_model=ProgramRead)
-def create_program(program_in: ProgramCreate, session: SessionDep):
+def create_program(program_in: ProgramCreate, session: SessionDep, current_user: Users = Depends(get_current_user)):
     program_data = program_in.model_dump(exclude={"days"})
     program = Program.model_validate(program_data)
+    program.owner_id = current_user.id
+
     session.add(program)
     session.flush()
 
@@ -65,11 +68,14 @@ def create_program(program_in: ProgramCreate, session: SessionDep):
 
 
 @router.put("/{program_id}", response_model=ProgramRead)
-def update_program(session: SessionDep, program_id: int, program_update: ProgramUpdate):
+def update_program(session: SessionDep, program_id: int, program_update: ProgramUpdate, current_user: Users = Depends(get_current_user)):
     program = session.get(Program, program_id)
 
     if not program:
         raise HTTPException(status_code=404, detail="Program not found")
+
+    if program.owner_id != current_user.id:
+            raise HTTPException(status_code=403, detail="Not enough permissions to edit this program")
 
     update_data = program_update.model_dump(exclude_unset=True, exclude={"days"})
     for key, value in update_data.items():
@@ -98,10 +104,14 @@ def get_program(session: SessionDep, program_id: int):
 
 
 @router.delete("/{program_id}")
-def delete_program(session: SessionDep, program_id: int):
+def delete_program(session: SessionDep, program_id: int, current_user: Users = Depends(get_current_user)):
     program = session.get(Program, program_id)
     if not program:
         raise HTTPException(status_code=404, detail="Program not found")
+    
+    if program.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not enough permissions to delete this program")
+
     session.delete(program)
     session.commit()
     return {"ok": True}
